refactor: extract terminal-task helpers to own module, soften chime comments
Codex second-pass review on the dedupe fix. All three items addressed:

1. Move `TERMINAL_STATUSES` and `tryMarkTaskTerminal` out of
   `packages/executor/src/index.ts` into a dedicated
   `terminal-task.ts` module. The helpers don't belong on the
   package's public surface, and the test no longer imports from the
   file that bootstraps the executor. Test renamed to match.

2. Soften over-claiming comments in `useTaskCompletionChime`. The
   previous text claimed the running-set dedupe handled "every known
   duplicate-event source" and that a stale seed ID "never triggers a
   chime" — both ignore the explicitly-accepted seed-vs-live RPC
   ordering edge case. Now the header documents the edge case as
   accepted scope and the seed comment cross-references it.

3. Drop the redundant standalone `already-COMPLETED` test in
   `terminal-task.test.ts` — it overlapped with the `it.each` block
   that already covers `COMPLETED`. Folded its distinguishing assertion
   (`tasks.get` was called) into the `it.each` block.
